Update fix_rigid_abrade.cpp
Updated setup_bodies_static() to only do the bare minimum setup of bodies. The compuation of the volume, mass, and inertia is now reserved for resetup_bodies_static() which is called after the initial neighbor list build. This removes the need to  preemtively call neighbor->build_topology() and forcing a neighborlist rebuild in pre_setup_neighbor().