Update fix_rigid_abrade.cpp
Removed unsupported optional keywords. Specifically, the mols keyword allowing for particles to be added on the fly. This functionality is unsupported since setting up the hybrid neighbour list during run times may be tricky. It is possible that the functionality can be regained if the hybrid neighborlist setup is moved from the input script to the source code in the future.